Enhance scripts to detect if we're missing mocha or jscoverage

+++ coverage.sh
@@ -1,6 +1,17 @@
 #!/bin/bash
 
-rm -rf ./coverage
-jscoverage ./src ./coverage
-mocha -R html-cov > ./test-coverage.html ./coverage
-open ./test-coverage.html
+if ((hash mocha 2>/dev/null) && (hash jscoverage 2>/dev/null)); then
+  echo "Clearing previous results"
+  rm -rf ./coverage
+
+  echo "Injecting jscoverage"
+  jscoverage ./src ./coverage
+
+  echo "Running tests (this may take a while)"
+  mocha -R html-cov > ./test-coverage.html ./coverage
+
+  echo "Displaying results"
+  open ./test-coverage.html
+else
+  echo "Please make sure Mocha and jscoverage are installed"
+fi
